requirements.txt
compress_json
encodeproject>=1.0.22
tqdm
ensmallen_graph>=0.4.3
node2vec
barplots
sanitize_ml_labels>=1.0.13
csrgraph
python-igraph
git+https://github.com/shenweichen/GraphEmbedding